rename pick to draft for legacy draft to be consistent with katowice drafting
